refactor(sera-types): unify ToolError, delete sera-runtime duplicate (sera-8s91/toolerr)
